[Feat] 그룹 게시글 분류 뷰 추가
